backend/requirements.txt
# Pinned for reproducibility; run pip-audit after updates (see AGENTS.md)
fastapi==0.135.3
python-multipart==0.0.20
uvicorn[standard]==0.32.1
python-dotenv==1.2.2
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6

# STEP 3 — LlamaIndex + Gemini (see AGENTS.md / backend/.env.example)
llama-index-core==0.14.20
llama-index-llms-google-genai==0.9.1
llama-index-embeddings-google-genai==0.5.0
google-genai==1.70.0

# tests / CI
pytest==8.3.4
httpx==0.28.1
pytest-asyncio==0.25.2
